Use restyle library
