+++ package.json
@@ -16,18 +16,19 @@
     "@react-native-masked-view/masked-view": "0.2.8",
     "@react-navigation/native": "^6.1.7",
     "@react-navigation/stack": "^6.3.17",
+    "@shopify/restyle": "^2.4.2",
     "axios": "^1.4.0",
+    "expo": "~48.0.18",
     "expo-blur": "~12.2.2",
     "expo-linking": "~4.0.1",
     "expo-status-bar": "~1.4.4",
     "expo-updates": "~0.16.4",
-    "expo": "~48.0.18",
+    "react": "18.2.0",
+    "react-native": "0.71.8",
     "react-native-gesture-handler": "~2.9.0",
     "react-native-reanimated": "~2.14.4",
     "react-native-safe-area-context": "4.5.0",
     "react-native-screens": "~3.20.0",
-    "react-native": "0.71.8",
-    "react": "18.2.0",
     "zustand": "^4.3.9"
   },
   "devDependencies": {
