+++ src/App.tsx
@@ -1,12 +1,14 @@
 import { StatusBar } from 'expo-status-bar'
 import { useMemo, useState } from 'react'
+import { Switch } from 'react-native'
 import { createStackNavigator } from '@react-navigation/stack'
 import { NavigationContainer } from '@react-navigation/native'
 import { SafeAreaView } from 'react-native-safe-area-context'
+import { ThemeProvider, createTheme } from '@shopify/restyle'
 
 import SearchScreen from './screens/Search.screen'
-import { createTheme, ThemeProvider } from './theme'
 import ResultsListScreen from './screens/ResultsList.screen'
+import { resolveThemeColorMode } from './theme'
 
 const Stack = createStackNavigator()
 
@@ -16,7 +18,7 @@ const THEME_PERSISTENCE_KEY = 'THEME_TYPE'
 export const App = () => {
   const [darkMode, toggleDarkMode] = useState(false)
 
-  const theme = useMemo(() => createTheme({ mode: darkMode ? 'dark' : 'light' }), [darkMode])
+  const theme = useMemo(() => resolveThemeColorMode(darkMode), [darkMode])
 
   return (
     <ThemeProvider theme={theme}>
