--- src/style-utils/README.md
@@ -1,3 +1,0 @@
-This is a customized "fork" of Shopify's Restyle library, which would work great out-of-the-box if not for the fact that I'm never satisfied with how colors are configured in any library's theme. I was pretty slap-dash with it, though.
-
-For the real deal, check out the real Restyle [here](https://github.com/Shopify/restyle). It's very nice.
