--- src/style-utils/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from './useResponsiveProp'
-export * from './useStyle'
-export * from './useTheme'
