--- src/style-utils/hooks/useResponsiveProp.ts
@@ -1,23 +1,0 @@
-import { useWindowDimensions } from 'react-native'
-
-import { PropValue, ResponsiveBaseTheme, ResponsiveValue } from '../types'
-import { getValueForScreenSize, isResponsiveObjectValue } from '../utilities'
-
-import { useTheme } from '@/theme/useTheme'
-
-const useResponsiveProp = <Theme extends ResponsiveBaseTheme, TVal extends PropValue>(
-  propValue: ResponsiveValue<TVal, Theme['breakpoints']>,
-) => {
-  const theme = useTheme()
-  const dimensions = useWindowDimensions()
-
-  return isResponsiveObjectValue(propValue, theme)
-    ? getValueForScreenSize({
-        responsiveValue: propValue,
-        breakpoints: theme.breakpoints,
-        dimensions,
-      })
-    : propValue
-}
-
-export { useResponsiveProp }
