--- src/style-utils/index.ts
@@ -1,12 +1,0 @@
-export * from './color'
-export * from './styleFunctions'
-export * from './types'
-export * from './createVariant'
-export * from './createBox'
-export * from './createText'
-
-export { useStyle } from './hooks/useStyle'
-export { useResponsiveProp } from './hooks/useResponsiveProp'
-export { createStyleFunction } from './createStyleFunction'
-export { composeStyleFunctions } from './composeStyleFunctions'
-export { createStyleComponent } from './createStyleComponent'
