--- src/style-utils/test/createText.test.tsx
@@ -1,157 +1,0 @@
-import React from 'react'
-import { create as render } from 'react-test-renderer'
-import { Text as RNText } from 'react-native'
-
-import { createTheme, ThemeProvider } from '../../theme'
-
-import { createText } from '../createText'
-
-const palette = {
-  black: '#000000',
-  grey: '#808080',
-}
-
-const theme = createTheme('light', {
-  colors: {
-    variables: {
-      black: palette.black,
-      grey: palette.grey,
-    },
-  },
-  spacing: {
-    s: 4,
-    m: 8,
-  },
-  textVariants: {
-    defaults: {
-      color: 'black',
-    },
-    title: {
-      fontSize: 48,
-      fontWeight: 'bold',
-      color: 'black',
-    },
-    subtitle: {
-      fontSize: 16,
-      color: 'grey',
-    },
-  },
-})
-
-type Theme = typeof theme
-// const textVariants = theme.textVariants.
-
-const Text = createText<Theme>()
-
-describe('createText', () => {
-  it('creates a bare Text component', () => {
-    const { root } = render(
-      <ThemeProvider theme={theme}>
-        <Text>Some text</Text>
-      </ThemeProvider>,
-    )
-
-    expect(root.findByType(RNText).props.children).toBe('Some text')
-  })
-
-  it('uses a color value from the theme', () => {
-    const { root } = render(
-      <ThemeProvider theme={theme}>
-        <Text color="black">Some text</Text>
-      </ThemeProvider>,
-    )
-
-    expect(root.findByType(RNText).props.style).toStrictEqual([
-      {
-        color: palette.black,
-      },
-    ])
-  })
-
-  it('uses a spacing value from the theme', () => {
-    const { root } = render(
-      <ThemeProvider theme={theme}>
-        <Text padding="2">Some text</Text>
-      </ThemeProvider>,
-    )
-
-    expect(root.findByType(RNText).props.style).toStrictEqual([
-      {
-        padding: 4,
-        color: palette.black,
-      },
-    ])
-  })
-
-  it('uses a text variant from the theme', () => {
-    const { root } = render(
-      <ThemeProvider theme={theme}>
-        <Text variant="title">Some text</Text>
-      </ThemeProvider>,
-    )
-
-    expect(root.findByType(RNText).props.style).toStrictEqual([
-      {
-        color: palette.black,
-        fontSize: 48,
-        fontWeight: 'bold',
-      },
-    ])
-  })
-
-  it('passes all React Native Text Style Props', () => {
-    const { root } = render(
-      <ThemeProvider theme={theme}>
-        <Text
-          color="grey"
-          fontFamily="Roboto"
-          fontSize={18}
-          fontStyle="italic"
-          fontWeight="800"
-          includeFontPadding
-          fontVariant={['lining-nums']}
-          letterSpacing={1.2}
-          lineHeight={1.4}
-          textAlign="left"
-          textAlignVertical="bottom"
-          textDecorationColor="black"
-          textDecorationLine="line-through"
-          textDecorationStyle="dashed"
-          textShadowColor="grey"
-          textShadowOffset={{ width: 4, height: 4 }}
-          textShadowRadius={16}
-          textTransform="capitalize"
-          verticalAlign="top"
-          writingDirection="rtl"
-        >
-          Some text
-        </Text>
-      </ThemeProvider>,
-    )
-
-    expect(root.findByType(RNText).props.style).toStrictEqual([
-      {
-        color: palette.grey,
-        fontFamily: 'Roboto',
-        fontSize: 18,
-        fontStyle: 'italic',
-        fontWeight: '800',
-        includeFontPadding: true,
-        fontVariant: ['lining-nums'],
-        letterSpacing: 1.2,
-        lineHeight: 1.4,
-        textAlign: 'left',
-        textAlignVertical: 'bottom',
-        textDecorationColor: palette.black,
-        textDecorationLine: 'line-through',
-        textDecorationStyle: 'dashed',
-        textShadowColor: palette.grey,
-        textShadowOffset: { width: 4, height: 4 },
-        textShadowRadius: 16,
-        textTransform: 'capitalize',
-        verticalAlign: 'top',
-        writingDirection: 'rtl',
-      },
-    ])
-  })
-})
