--- src/style-utils/utilities/index.ts
@@ -1,4 +1,0 @@
-export { getValueForScreenSize } from './getValueForScreenSize'
-export { getResponsiveValue } from './getResponsiveValue'
-export { isResponsiveObjectValue } from './isResponsiveObjectValue'
-export { getThemeValue } from './getThemeValue'
