+++ src/theme/color/index.ts
@@ -1,4 +1,5 @@
 export * from './palette'
 export * from './color.base'
+export * from './color.validators'
 export * from './color.variants'
 export * from './color.types'
