--- src/theme/useTheme.ts
@@ -1,8 +1,0 @@
-import { useContext } from 'react'
-
-import { Theme } from './createTheme'
-import { ThemeContext } from './theme.context'
-
-const useTheme = () => useContext(ThemeContext) as Theme
-
-export { useTheme }
